% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_circos_links.R
\name{color_circos_links}
\alias{color_circos_links}
\title{color_circos_links}
\usage{
color_circos_links(df, hue = c(" ", "random", "red", "orange", "yellow",
  "green", "blue", "purple", "pink", "monochrome"), luminosity = c(" ",
  "random", "light", "bright", "dark"))
}
\arguments{
\item{df}{A data frame containing circos links information - as from `make_circos_links()`.}

\item{hue}{Desired hue of the links.}

\item{luminosity}{Desired luminosity of the links.}
}
\value{
Returns original links data frame with an added `$color` column in RGB format for Circos plotting.
}
\description{
A function designed to assign random colors from the `randomcoloR` package to a `make_circos_links()`
generated data frame.
}
\examples{
df <- data.frame(x = rnorm(50), y = rnorm(50))
str(df)
df <- color_circos_links(df)
str(df)
}
\author{
Matthew Woodruff, Emory University
}
