% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_searchable_links.R
\name{generate_searchable_links}
\alias{generate_searchable_links}
\title{generate_searchable_links}
\usage{
generate_searchable_links(df, chromosome_grouping, band_grouping, link_grouping,
  start_position, end_position)
}
\arguments{
\item{df}{A data frame containing all sub-lineages from the same lineageID}

\item{chromosome_grouping}{A column defining the circos chromosomes. Passed as column index or "column name".}

\item{band_grouping}{A column defining the circos chromosome bands. Passed as column index or "column name".}

\item{link_grouping}{A column defining the circos linkage IDs. Passed as column index or "column name".}

\item{start_position}{A column defining the start positions of the linkage ID.}

\item{end_position}{A column defining the end positions of the linkage ID.}
}
\value{
Returns a data frame in a pre-"circos links" format for later use within the `make_circos_links` function.
}
\description{
A function serving as the core of the make_circos_links function. Takes a data frame containing a single lineage,
and generates a links data frame containing all possible combinations of that lineage.
}
\author{
Matthew Woodruff, Emory University
}
